+++ www/whatsnew.md
@@ -17,6 +17,7 @@ This release includes bug fixes reported by the community.
  - Darius Crowley no longer survives lethal damage from its effect. Student of the Ox, which has an effect with a similar consequence, **does** survive lethal damage from its effect on purpose.
  - Shudderwock and other random card casting effects now stop casting when they are removed from play due to a transform effect, like Thrall Deathseer.
  - Ugar Frostflame, Deepholm Portal, and Fists of Fury no longer prompt the player for targets.
+ - Fifi Fizzlewarp no longer writes a Permanent's text on a minion.
 
 ###### 0.7.7-2.0.12 (Saturday, December 8th)